integration

+++ Quantizer.py
@@ -7,7 +7,7 @@ class Quantizer:
     @staticmethod
     def Quantize(input_vector: np.ndarray) -> np.ndarray:
         quantize_matrix = np.ones(input_vector.shape) * Quantizer.quantize_step
-        return np.sign(input_vector) * np.fix(np.abs(input_vector) / quantize_matrix)
+        return (np.sign(input_vector) * np.fix(np.abs(input_vector) / quantize_matrix)).astype(np.int64)
 
     @staticmethod
     def dequantize(input_vector: np.ndarray) -> np.ndarray:
